Move `CommandTransactionBuilder`
Move the `CommandTransactionBuilder` to a new
families::command::transaction_builder module which is behind the
"family-command-transaction-builder" feature. Moving this builder
to the new "family-command-transaction-builder" feature allows it
to be used without having to pull in the additional workload
infrastructure that is behind the "family-command-workload"
feature.
